+++ test/decorators-duplicate-fixture-name/playwright.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from '@playwright/test';
+import { defineBddConfig } from '../../dist';
+
+const testDir = defineBddConfig({
+  importTestFrom: 'fixtures.ts',
+  paths: [`sample.feature`],
+});
+
+export default defineConfig({
+  testDir,
+  forbidOnly: Boolean(process.env.FORBID_ONLY),
+});